# project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SMM Project API",
        "description": "API for SMM project management system",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/v1",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "Retrieve paginated list of projects with tasks, members, invitations and chats",
                "operationId": "9d1f91c9d91de620a0d5f11b734d7bef",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Project"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "description": "Create a new project with the provided data",
                "operationId": "4d679b7e3456e4d811314aaebb5ae51d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Marketing Campaign"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Social media marketing campaign for Q4"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "completed",
                                            "on_hold"
                                        ],
                                        "example": "active"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-21"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "budget": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get project details",
                "description": "Retrieve detailed information about a specific project including tasks, members, invitations and chats",
                "operationId": "a3902ae34b85515b32e4108296c3d4e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update project",
                "description": "Update an existing project",
                "operationId": "9abcb3f9742b94aa18e5c4d0ca106f3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Marketing Campaign"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "completed",
                                            "on_hold"
                                        ],
                                        "example": "active"
                                    },
                                    "budget": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete project",
                "description": "Delete an existing project",
                "operationId": "5dd94942bc2f64f2d041f40ac39dce87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Project deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Project deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/projects/{id}/complete": {
            "patch": {
                "tags": [
                    "Projects"
                ],
                "summary": "Mark project as completed",
                "description": "Change project status to completed",
                "operationId": "45d08a5df8dc6cab2c5ab627cdaa311f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project marked as completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Project": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Marketing Campaign Q4"
                    },
                    "description": {
                        "type": "string",
                        "example": "Social media marketing campaign for Q4 2025"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "completed",
                            "on_hold"
                        ],
                        "example": "active"
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-21"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "budget": {
                        "type": "number",
                        "format": "float",
                        "example": 15000.5
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_completed": {
                        "type": "boolean",
                        "example": false
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Task"
                        }
                    },
                    "members": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectMember"
                        }
                    },
                    "invitations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectInvitation"
                        }
                    },
                    "chats": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Chat"
                        }
                    },
                    "stats": {
                        "$ref": "#/components/schemas/ProjectStats"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-21T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-21T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Task": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Create social media posts"
                    },
                    "description": {
                        "type": "string",
                        "example": "Design and create posts for Instagram and Facebook"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed",
                            "cancelled"
                        ],
                        "example": "pending"
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ],
                        "example": "high"
                    },
                    "assigned_to": {
                        "type": "integer",
                        "example": 2
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-30"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-21T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "ProjectMember": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "owner",
                            "manager",
                            "member",
                            "viewer"
                        ],
                        "example": "manager"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "read",
                            "write",
                            "manage_members"
                        ]
                    },
                    "joined_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-21T10:00:00Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "ProjectInvitation": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "manager",
                            "member",
                            "viewer"
                        ],
                        "example": "member"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "accepted",
                            "declined",
                            "expired"
                        ],
                        "example": "pending"
                    },
                    "token": {
                        "type": "string",
                        "example": "abc123..."
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    },
                    "invited_by_user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Chat": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "message": {
                        "type": "string",
                        "example": "Hello team! Let's discuss the project requirements."
                    },
                    "message_type": {
                        "type": "string",
                        "enum": [
                            "text",
                            "voice",
                            "file"
                        ],
                        "example": "text"
                    },
                    "sender_type": {
                        "type": "string",
                        "enum": [
                            "customer",
                            "admin"
                        ],
                        "example": "customer"
                    },
                    "file_path": {
                        "type": "string",
                        "example": "http://localhost/storage/files/document.pdf"
                    },
                    "is_read": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_voice_message": {
                        "type": "boolean",
                        "example": false
                    },
                    "customer": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-21T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "http://localhost/storage/avatars/avatar.jpg"
                    },
                    "role": {
                        "type": "string",
                        "example": "project_manager"
                    }
                },
                "type": "object"
            },
            "ProjectStats": {
                "properties": {
                    "total_tasks": {
                        "type": "integer",
                        "example": 15
                    },
                    "completed_tasks": {
                        "type": "integer",
                        "example": 8
                    },
                    "pending_tasks": {
                        "type": "integer",
                        "example": 5
                    },
                    "overdue_tasks": {
                        "type": "integer",
                        "example": 2
                    },
                    "completion_percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 53.33
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}